commit de emergencia

+++ Holy_Graph/Cursus42/libft/functions/ft_strlen.c
@@ -3,15 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: thfirmin <thiagofirmino2001@gmail.com>     +#+  +:+       +#+        */
+/*   By: thfirmin <thfirmin@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/05/05 03:39:57 by thfirmin          #+#    #+#             */
-/*   Updated: 2022/05/11 03:05:35 by thfirmin         ###   ########.fr       */
+/*   Created: 2022/05/12 20:35:17 by thfirmin          #+#    #+#             */
+/*   Updated: 2022/05/12 23:51:16 by thfirmin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-//calculate the length of a string
-unsigned int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
 	unsigned int	count;
